Move functions to dump a generic json resource to its own file
